+++ Tunny/WPF/ViewModels/Visualize/RankViewModel.cs
@@ -1,11 +1,21 @@
 ﻿using System.Linq;
 
+using Optuna.Study;
+
 using Tunny.Core.Settings;
 
 namespace Tunny.WPF.ViewModels.Visualize
 {
-    class RankViewModel : PlotSettingsViewModelBase
+    internal class RankViewModel : PlotSettingsViewModelBase
     {
+        public RankViewModel() : base()
+        {
+        }
+
+        public RankViewModel(StudySummary[] summaries) : base(summaries)
+        {
+        }
+
         public override PlotSettings GetPlotSettings()
         {
             return new PlotSettings