+++ Tunny/WPF/ViewModels/Visualize/SliceViewModel.cs
@@ -1,11 +1,21 @@
 ﻿using System.Linq;
 
+using Optuna.Study;
+
 using Tunny.Core.Settings;
 
 namespace Tunny.WPF.ViewModels.Visualize
 {
-    class SliceViewModel : PlotSettingsViewModelBase
+    internal class SliceViewModel : PlotSettingsViewModelBase
     {
+        public SliceViewModel() : base()
+        {
+        }
+
+        public SliceViewModel(StudySummary[] summaries) : base(summaries)
+        {
+        }
+
         public override PlotSettings GetPlotSettings()
         {
             return new PlotSettings()