+++ Tunny/WPF/Views/Pages/Visualize/HypervolumePage.xaml.cs
@@ -1,4 +1,8 @@
 ﻿using System.Windows.Controls;
+
+using Optuna.Study;
+
+using Tunny.WPF.ViewModels.Visualize;
 namespace Tunny.WPF.Views.Pages.Visualize
 {
     public partial class HypervolumePage : Page
@@ -6,6 +10,12 @@ namespace Tunny.WPF.Views.Pages.Visualize
         public HypervolumePage()
         {
             InitializeComponent();
+        }
+
+        public HypervolumePage(StudySummary[] summaries)
+        {
+            InitializeComponent();
+            DataContext = new HypervolumeViewModel(summaries);
         }
     }
 }