+++ Tunny/WPF/Views/Pages/Visualize/TimelinePage.xaml.cs
@@ -1,5 +1,9 @@
 ﻿using System.Windows.Controls;
 
+using Optuna.Study;
+
+using Tunny.WPF.ViewModels.Visualize;
+
 namespace Tunny.WPF.Views.Pages.Visualize
 {
     public partial class TimelinePage : Page
@@ -7,6 +11,12 @@ namespace Tunny.WPF.Views.Pages.Visualize
         public TimelinePage()
         {
             InitializeComponent();
+        }
+
+        public TimelinePage(StudySummary[] summaries)
+        {
+            InitializeComponent();
+            DataContext = new TimelineViewModel(summaries);
         }
     }
 }